requirements.txt
# MoodWave - AI Music Curation & Event Discovery Engine
# Core
pandas>=2.0
numpy>=1.24
pydantic>=2.0
pydantic-settings>=2.0

# ML / Embeddings
sentence-transformers>=2.2
faiss-cpu>=1.7

# API
fastapi>=0.100
uvicorn[standard]>=0.20
httpx>=0.24
python-dotenv>=1.0

# Existing
PyPDF2>=3.0
